requirements.txt
openai
requests
python-dotenv
feedparser
yt-dlp
ffmpeg-python
whisper
